refactor(gateway): extract start_server and route composition into platform/router.rs — ironclaw#2599 stage 2
Second increment of the ironclaw#2599 platform/feature split. Moves
`start_server()` and the Axum route composition out of `server.rs` into
a dedicated `platform/router.rs`, so the platform-vs-features
dependency direction is visible: the router depends on handler modules
(both `handlers/*` and the still-inline handlers in `server.rs`),
never the reverse.

Changes:

- New `src/channels/web/platform/router.rs` owns `start_server()`, the
  four routers (`public`, `protected`, `statics`, `projects`), and the
  cross-cutting layer stack (CORS, 10 MB body limit, panic catch,
  `X-Content-Type-Options`, `X-Frame-Options`, CSP).
- Feature handlers still inline in `server.rs` are now `pub(crate)` so
  the router can register them without leaking them outside the crate.
  Two private structs that are directly referenced by pub(crate)
  handlers (`HistoryQuery`, `GatewayStatusResponse`) were also raised
  to `pub(crate)` so the handler signatures type-check from the router
  module.
- `server.rs` keeps the feature handlers that haven't migrated yet
  (OAuth callbacks, chat, extensions, pairing, logs, gateway status)
  and adds `pub use platform::router::start_server` so external call
  sites — `src/channels/web/mod.rs`, `tests/multi_tenant_integration.rs`
  — keep working. The trimmed imports drop `Router`, `DefaultBodyLimit`,
  `CorsLayer`, `tokio::sync::mpsc`, etc., since they're no longer used
  in the remaining body.
- `mod.rs` now calls `platform::router::start_server` directly; the
  `server::start_server` shim exists only for external code paths that
  still reach for it.
- `CLAUDE.md` file map now lists `platform/router.rs` and clarifies
  that `server.rs` is feature-handler-only pending migration.

No behavior change. Route table, middleware stack, CORS policy, body
limits, panic handling, security headers, and CSP are byte-identical
to origin/staging.

Stats: server.rs 7463 → 6973 lines (−490); new `platform/router.rs` is
537 lines. `cargo clippy --all --benches --tests --examples
--all-features` is clean. Unit test run: 5068 passed; the same 2
pre-existing failures carried over from stage 1
(`pairing::approval::tests::propagate_approval_restores_runtime_state_when_on_start_fails`
needs a telegram WASM fixture;
`extensions::manager::tests::test_telegram_token_colon_preserved_in_validation_url`
has a test-infra URL override — neither references any symbol this
PR touches).
